django-app/pyproject.toml
[tool.poetry]
name = "redbox_app"
version = "0.1.0"
description = ""
authors = ["i.AI <i-dot-ai-enquiries@cabinetoffice.gov.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "^5.0.7"
django-libsass = "^0.9"
django-magic-link = "^1.0.0"
whitenoise = "^6.6.0"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
jinja2 = "^3.1.2"
django-environ = "^0.11.2"
django-use-email-as-username = "^1.4.0"
markdown-it-py = "^3.0.0"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
pytz = "^2024.1"
django-permissions-policy = "^4.18.0"
django-csp = "^3.7"
django-single-session = "^0.2.0"
sentry-sdk = {extras = ["django"], version = "^2.5.0"}
django-storages = "^1.14.2"
boto3 = "^1.34.65"
urllib3 = "<2.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"
yarl = "^1.9.4"
django-allauth = "^0.61.1"
humanize = "^4.9.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
django-gov-notify = "^0.5.0"
websockets = "^12.0"
django-import-export = "^4.0"
dataclasses-json = ">=0.6.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-django = "^4.8.0"
pytest-env = "^1.1.1"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23"
pylint = "^3.0.3"
mypy = "^1.9.0"
types-python-dateutil = "^2.9.0.20240315"
types-markdown = "^3.6.0.20240316"
ruff = "^0.3.3"
python-dotenv = "^1.0.1"
pre-commit = "^3.6.2"
freezegun = "^1.2.2"
bandit = "^1.7.8"
django-types = "^0.19.1"
types-python-jose = "^3.3.4.8"
detect-secrets = "^1.4.0"
types-pytz = "^2024.1.0.20240203"
boto3-stubs = "^1.34.67"
requests-mock = "^1.12.1"
playwright = "^1.43.0"
pytest-playwright = "^0.4.4"
axe-playwright-python = "^0.1.4"
django-test-migrations = "^1.3.0"
beautifulsoup4 = "^4.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "redbox_app.settings"
